// interfaces/Favorites.ts
import { Artist } from './Artist';
import { Album } from './Album';
import { Track } from './Track';

export interface Favorites {
  artists: string[]; // favorite artists ids
  albums: string[]; // favorite albums ids
  tracks: string[]; // favorite tracks ids
}

export interface FavoritesResponse {
  artists: Artist[];
  albums: Album[];
  tracks: Track[];
}